Remove obsolete file
